* Added `literal-style?' and `content-type' keyword parameters for `html-page'.
* The `content-type' and `charset' keyword parameters for `html-page'
work together: if `content-type' is provided and `charset' is not
provided, `charset' is assumed to be `"UTF-8"'; if `charset' is
provided and `content-type' is not provided, `content-type' is assumed
to be `"text/html"' (if html-tags's `generate-sxml?' is `#f') or
`"application/xhtml+xml"' (if `generate-sxml?' is non-`#f').

* Warning: the default behavior for inlined CSS (`css' keyword
parameter for `define-page') has changed.  Now the special HTML
characters are converted to their corresponding entities in non-SXML
mode (i.e., when `generate-sxml?' is `#f' -- the default value).  This
may break your page styles if you inline CSS via `html-page''s `css'
keyword parameter.  To get the old behavior, use `literal-style?: #t'
for `define-page'.